Ignore invalid usage of TraceState instead of throwing.
